requirements.txt
Flask
python-dotenv
schedule
twilio
newsapi-python